:recycle: [Refactor] 파일 이동 및 불필요 파일 삭제
