--- src/main/java/samryong/config/SwaggerConfig.java
@@ -1,23 +1,0 @@
-package samryong.config;
-
-import io.swagger.v3.oas.models.Components;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-
-    @Bean
-    public OpenAPI openAPI() {
-        return new OpenAPI().components(new Components()).info(apiInfo());
-    }
-
-    private Info apiInfo() {
-        return new Info()
-                .title("Samryong BackEnd API")
-                .description("Samryong BackEnd API 명세서")
-                .version("1.0.0");
-    }
-}
