+++ src/main/java/samryong/global/config/WebConfig.java
@@ -1,0 +1,20 @@
+package samryong.global.config;
+
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import samryong.security.resolver.GetTokenArgumentResolver;
+
+@RequiredArgsConstructor
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+
+    private final GetTokenArgumentResolver getTokenArgumentResolver;
+
+    @Override
+    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
+        resolvers.add(getTokenArgumentResolver);
+    }
+}
